ports: event_out_port: fix order of args in is_constructible and don't remove_reference.
The arguments to is_constructible should be type first, then args
to attempt to construct from. Due to the interaction of universal refs
with remove_reference, lvalue references silently made it through
the static assert until they are static_casted to event_t.

So a caller could accidentally pass an lvalue to fire and then get it
moved from if event_t is an rvalue ref. This now needs to be an explicit
decision (such as moving the args to fire).