core: disallow abusing active_connection_proxy.
By storing an active_connection_proxy in a variable and using it
multiple times it was possible to do evil things. This is disallowed by
only giving active_connection_proxy rvalue overloads of .connect(). Now
either the ACP is used in it's natural way or the user must consciously
cast it to an rvalue.