logging: add first implementation.
Currently supports adding file, syslog and stream sinks, and getting a
pointer to a static logger instance. The stream sink adding function
returns a handle that removes the stream from the logger when the handle
is destroyed. This is supposed to ensure that local streams are not used
after return.