node_aware: FIX regression in connecting node_aware to mixed node_aware.
Previously there was a direct check of is_instantiation_of<node_aware>
with the passive side of the connection, which didn't catch node_awares
that are wrapped in other mixins. This lead to no buffers being
introduced. This wasn't previously visible in tests but can be seen when
running dummy_main.

The new approach is to check whether an object of the passive type binds
to a `const node_aware<base>&` where the type of base is deduced. This
should catch all cases. The buffers are now correctly introduced in our
test cases.

The test case had to be fixed too because it didn't provide
specializations of is_active_sink/is_active_source for useless_mixin.