node_aware: don't store introduce_buffer() output in temporary variable, use directly.
The buffered connection object is a local lvalue and so it needs to be
moved in the call to base::connect. The previous version used
std::forward which does an unconditional move in this case, which was
confusing.

Now the return value of the introduce_buffer() call is directly used in
the call to base::connect which prevents confusion.