graph: forward declare graph_connectable and move graph_adder definition up.
graph_adder is used in graph_connectable but it is defined after. This
conflicts with two-phase name lookup, as graph_adder is not a dependent
name of the template parameters of graph_connectable.

graph_connectable then needs to be forward declared before graph_adder,
and is looked up in the second phase.

Bug caught by clang.