FIX generalized lambda capture of unique_ptr.
In C++14 `[x{y}]` is equivalent to `auto x{y};` which deduces the type
of x to `std::initializer_list<decltype(x)>`. This a language defect for
which a fix was described in proposal N3922[1], which was accepted.
This was implemented by gcc in c++14 mode but not in clang until version
3.8. Due to this it's best to use the unambigous version without a
braced-init-list.

Additionally moving the result of make_unique is redundant - the return
value of make_unique can already be used as an rvalue.

[1]: http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2014/n3922.html