periodic_task: remove indirection in work_to_do and add heap allocated mutex.
The shared_ptr<bool> is an artifact of an era long gone, when ports
and periodic tasks used to be copyable. The real problem this commit
solves is that the work_to_do bool was being read/written from multiple
threads:

  * cycle_control main thread checks whether the task is done.
  * cycle_control main thread sets it to true if the task is due.
  * the worker thread sets it to false when it completes the work.

I complete redesign of the scheduler is probably necessary but guarding
the bool with a mutex is needed until then.