dummy_main: manually call cycle_control::stop().
Since the scheduler is created as the first object in main, it will be
destroyed last. It has access to tasks created after it, meaning that
when the destructors run in the normal order, the scheduler will still
be executing tasks while all connectables are being destroyed.

To prevent this data race, manually call cycle_control::stop() before
main() exits.