dummy_main: check for exceptions and return from main if out_of_time.
Ignoring the exception leads to the main scheduler loop exiting, and the
worker threads waiting for more work. The owner of the scheduler needs
to check the exception to know whether the loop is still running, and if
there is an exception the only solution is to log and exit.