ports: buffers: introduce third buffer and additional switch tick.
Previously the buffer was connected to a switch tick from one region and
a work tick from the other. Since switch ticks and work ticks are not
synchronized between regions, this lead a race conditions on the buffer.
One thread could be reading from the buffer while another was trying to
switch the buffers.

After discussions with Caspar, we decided against adding a mutex to the
buffer. While this would prevent the race condition, it would still not
give us the order guarantees that we need to make the signal flow
deterministic.

Instead we came up with the following triple buffered scheme:

          -----  -----  -----
     WT1  |   |  |   |  |   |  WT2
    ----->| I |  | M |  | O |----->
          |   |  |   |  |   |
          -----  -----  -----
            \____^   \____^
             ST1      ST2

With this solution switch ticks from two regions need to happen to get
data from the incoming buffer (I) through the middle buffer (M) to the
outgoing buffer (O). Because the switch ticks are strictly sequential
(they run in the main scheduler loop) and they run after the region's
worker thread is done, we can be *reasonably* sure that there is no race
condition in the buffers.

The two switch ticks are called switch_active_tick and switch_passive_tick.
They need to be connected on the active and passive node sides accordingly.

When using buffers directly this means their usage is asymmetric
depending on whether we're dealing with events or states. For states
the order should be:

  1. work_tick
  2. switch_passive_tick
  3. switch_active_tick

and for events:

  1. switch active_tick
  2. switch_passive_tick
  3. work_tick

This difference comes down to the direction of the data flow.