cycle_control: disallow adding tasks to running cycle_control.
This is unsafe because the tasks are stored in a vector and the worker
threads only have a reference to the task. If a task is added and
the vector needs to be reallocated then all references are invalidated.

This commit adds a check whether the scheduler has been started, and
throws if an attempt is made to a add_task to a running cycle_control.
This is the safest solution and corresponds to the most common case.
This does not affect the scheduler itself - we add tasks to it all the
time.

With this, the task_queue mutex is not needed anymore.