parallel_scheduler: FIX main_loop of worker threads to prevent data races on do_work.
do_work and the task_queue are both protected by the task_queue_mutex.
This means do_work cannot be read while the mutex is not held.
Additionally, spurious wakeups are always possible and a condition
variable which is used to signal threads that "the task queue is not
empty" needs to checked after a wakeup.