threading: remove tick_rate from periodic_task.
The tick length now needs to be passed to cycle_control along with the
periodic_task in the call to cycle_control::add_task(). This will allow
the cycle_control to split up tasks into separate queue's for various
tick rate's. With this it should be possible to make stronger guarantees
on ordering of tasks with different cycles.

In particular, when using separate queues, we can make sure that every
10'th fast tick will happen before the next medium tick.

For the time cycle_control has no way of knowing when tasks are due
which means that all tasks will be run with a period of min_tick_length.