cycle_control: wait until now+min_tick in main_loop.
This should add up to the correct amount of waits over all loop
iterations, and the wait until now+max_tick afterwards is not needed.

It's better to advance the virtual clock after the main loop, that way
all tasks start at the first clock. In the long run it doesn't make a
difference but for the dummy_main test the advance after the work is
what we want.