event_sink: implement registering callbacks directly in class.
The base class wasn't a real base class; it existed only to ease
implementation. Composition should be preferred to inheritance, so the
required functionality was implemented directly in event_sink. Once this
is needed in state_source most of it can be copied, or factored out into
a common object.

Scoped deregister is also not needed since ports are noncopyable.