ports: move connection_breaker to port_utils and add policy.
The policy is used to distinguish between ports that can have a single
handler and those that can have multiple handlers. This is needed to
allow connection_breaker to work with both event_source and state_sink.

There is an interesting corner case in single_handler_policy when the
handler is overridden. The previous handler still has a registered
callback and will try to remove itself on destruction. We check whether
the handler that is being removed is the same one as the _last_ one that
we stored. This works but it remains to be checked if this is the best
approach.