ports: mux: add merging functionality.
Merging is based on two classes: loaded_merge_port and
unloaded_merge_port. When merge() is called with a merge function an
unloaded merge_port is created. This port can not be connected to
anything but a mux_port.

Once the unloaded_merge_port reaches a mux_port a loaded_merge_port is
created, which is a passive source. It can be further connected until an
active sink is reached.