infrastructure: capture region in periodic_task.
Without this the periodic_task relies on someone else keeping the region
alive. This is dangerous - as mentioned in the comment it can easily lead
to use-after-free as shown by the following code:

    fc::infrastructure infra;
    infra.add_region(...);
    infra.start_scheduler();

The region is a shared entity any way, so storing it is explicitly
allowed. This doesn't lead to a reference cycle because the region is
the passive side of the connection with the periodic_task.