infrastructure: stop the scheduler in destructor.
The scheduler needs to be stopped before the connectables start getting
destroyed. We found out about this the hard way in dummy_main. The
solution was to manually stop it on every exit path (with a scope_exit
style macro).

With the new infrastructure, ideally we would like to order the members

  1. root_node
  2. scheduler

but can't because the root_node needs a region and our add_region method
directly hooks the region up to the scheduler. What we can do, and do
after this patch, is stop the scheduler in the destructor. This happens
before the root_node is destroyed and means that typical users of
infastructure do not have to worry about stopping the scheduler
manually.

Unfortunately, this is not enough for  dummy_main because it mixes nodes
with manually managing regions and ports. There we still have to stop
the scheduler before the exit from main.