cmake: Add install instructions
The install section installs headers (including 3rdparty), the built
flexcore lib and cmake -config/-targets files. We also deal with
multilib installations.

The only correct way of doing this is to put 64-bit libs in
<prefix>/lib64 and 32-bit libs in <prefix>/lib32. Debian handles this
differently and puts them in <prefix>/lib/<arch-tuple>. In order to not
close this door, the flag LIB_SUFFIX can be specified in the cmake
invocation as something like "/x86_64-linux-gnu".