docs: Add GETTING_STARTED and adjust doxyfile
The main page is now correctly shown in doxygen output, it links to the
getting started page, and that includes a WIP description of basic
flexcore building blocks. The documentation files now have the md suffix
to invoke the correct parser. Test and build directories are excluded
from doxygen output.