nodes: Remove self iterator from tree_base_node
The iterator to self in forest can be found using a ptr to forest and
this. This was causing more grief than good because it couldn't be set
in the constructor which caused problems when writing compound nodes.

At the moment tree_base_nodes have no way of accessing the forest which
means they cannot access their fullname. When we make the connection
graph a part of the forest we will need to have a way of accessing
atleast the root node of the tree.