nodes: Add forest* member to tree_base_node
If graph is supposed to be non-global, then it needs to be tied to the
forest in some way. Therefore nodes need to be able to access the
forest.

There were very many tests that relied on the specific signature of the
tree_base_node constructor. We should look into reducing coupling on
this.