nodes: Add graph to root_node
With this the graph is not a global state anymore. The graph is placed
inside the root_node, althought this is probably a temporary measure.
That way it can be accessed through the forest, and the forest ptr
placed inside tree_base_node.

An alternative would be to have only a graph pointer in tree_base_node
and store ptr to forest in owning_base_node. This way the graph could be
held by infrastructure. This would restrict access to the forest.

These changes will be made at a later time.