nodes: Pass tree_base_node instance to child constructors
Previously all nodes that wanted to be part of tree needed to derive
from tree_base_node and know the details of how to construct one. Now
derived classes only need to copy a "protypical" tree_base_node. The
default name is now pulled from a static member called 'default_name'.

This hides most of the details of tree_base_node from users.

An alternative would be to pass some placeholder object ('tree_args')
that a tree_base_node can be constructed from. This is effectively the
same.