scheduler: Add new_region() method to parallel_region
For normal parallel_regions this is equivalent to the constructor. But
for infrastructure created regions, we will be able to override
new_region() to provide a region that is hooked up to the scheduler.

I did not give parallel_region a virtual destructor so that people don't
get the idea of using them polymorphically. Since we deal with
shared_ptr<parallel_region> in flexcore, this is not a problem (the
correct destructor will be called for a derived class).