pure_ports: Store handlers together with their hashes
The previous design, where the handlers where part of the port, the
hashes where in a different object and the conenction breaker callback
needed to access both was nonesense. It lead to internal pointers
between members that could not be updated at the same time and made move
constructors impossible to implement.

Now the old 'connection_breaker' class is renamed to 'active_port_base'
which is a much better name for it. It is a slightly more sophisticated
storage for handlers and their hashes, that registers the callback that
disconnects ports when they are destroyed.

This commit pretty much only moves the handlers into the
active_port_base, it does not fix the moving bugs yet.