pure_ports: Reconstruct callback in move constructor
A callback, as the name suggests, needs to have a way to call back to
an object. We do this by giving it a copy of the 'this' ptr. When an
object is moved, its (abstract) identity is unchanged but the value of
'this' changes.

Previously I forgot about this and so the whole connection breaker stuff
didn't work when an active port was moved, because the callback
contained a dangling pointer. Now this is not an issue anymore.