pure_ports: single_handler: Swap in move ctr as a workaround for libc++
Libc++ copies the function object in the move constructor. I don't think
this is _wrong_ per se, but its different to what libstdc++ does. As far
as I can tell even more recent version of libc++ do this (checked 3.8).

To keep things consistent we perform a swap, so that the
moved-from-object gets replaced with a default constructed (empty) std
function. With this a moved from port can not access its previous
handler after a move, which is the desired behaviour.