pure_ports: Assert when connected passive ports are moved
This is undesirable because an active port keeps a reference to the
passive port. There is no way to modify this reference later, which
would be dangling after a move, so we need to assert.

An interesting thing here is that when the assert is compiled out
(NDEBUG) the moved-from-object will be gracefully disconnected from it's
active ports. This happens thanks to us not moving the connection
breakers.