traits: Use source information in new is_passive_sink trait
It's very hard to check wheter a generic lambda is a passive sink. This
led connections like this:

    event_source<int> x;
    x >> [](auto) {};

to not result in a port_connection but instead in an
active_connection_proxy. This meant that when x.fire() was called
nothing happened, because the connection was never fully established.

This came down to the checks for is_passive_sink not working for generic
lambdas. But if we use the information on the source part, we can check
if the right-hand side behaves like a proper sink. This is what the new
'is_passive_sink_for' trait does.

This trait is now used in active_passive_connect_impl and
active_connection_proxy.