tests: Test that chaining generic lambdas in event connections works
Generic lambdas are tricky to get right because there is no way to
introspect at compile time whether they have a call operator. The only
way this can be done is if we know the argument types.

It so happens that when building a connection, in most cases the
argument type is known. That's what the new 'is_passive_sink_for' trait
does. This test checks that it works in one of the most pathological
cases: a connection with 3 generic lambdas, where the last one is a
sink.