nodes: Make everything in node_args private and add friends
This way our 3 classes that do need access to node_args (tree_base_node,
owning_base_node, forest_owner) can do what they please while client
code can't inspect node_args. This was the desired behaviour from the
beginning, it just took us a bit longer to figure out how to achieve
this goal.

A lot of this commit is just adapting the tests to this new behaviour.
I did not want to special case test code so instead I abused inheritance
and knowledge of forest internals to minimize the amount of test code
that needs to be changed.