tests: Fix fast_main_loop test to not fail on slow machines
Previously the test would let tasks run for 10ms, with the assumption
that this is enough time for slow tasks to run atleast once. On slow
machines (such as some jenkins build nodes) this is not true.

Now a std::promise is used to block the test thread until the slow task
has had the chance to run. I tested that this test doesn't fail anymore
by running (on an 8 core machien):

    ( one terminal )
    stress --cpu 16

    ( other terminal)
    for i in `seq 1 100`; do
        ./tests/test_executable &
    done
    wait