tests: Fix data races in scheduler related tests
Thread sanitizer reported some 5-15 races, most of them related to the
destructor of std::promise/std::future. It appears that it is not
correct to have a promise object shared between two threads, even if one
of them is supposed to block on future::get(), because it may be
released and race to the destructor before promise::set_value()
finishes.

These uses have been replaced with atomics and explicit yielding.