Enhancement to compiler errors on invalid ctor calls of passive_ports
ctr of event_sink and state_source now take action by universal reference.
They then convert to std::function internally.
This creates slightly better error messages, as the failure to convert the action to the std::function
raises the error and not the missing ctor of the port.

Also added static asserts to the bodies of these ports.
These asserts only cause the 2nd or 3rd error message in compilation,
but hint to the mistake the programmer made.