scripts: Fix processing script to handle compound nodes with ports
The previous script worked for nodes with ports (connections) and
pure compound nodes. It couldn't handle the case of compound nodes that
also contained ports, because that dosen't really map to graphviz.

So that we get _any_ output for such cases, I hacked around the
limitations of graphviz by rendering compound nodes with ports as a
subgraph that contains a node by the same name.