scheduler: Fix cycle_control::stop to not join thread pool
Joining the threads meant that it was not possible to restart the
scheduler, because it did not own the threads that it stopped. Instead
we now join the scheduling thread and wait for the tasks to be done,
giving them up to slow_tick to finish. This means that the thread pool
is alive and scheduling tasks can be resumed later.

Fixes #22514.