periodic_task: Explicitly store region inside periodic task
Previously, regions were captured by the lambda that is used to
construct the std::function object. Now they are explicitly captured
inside the periodic task. This will make it possible to access the
region a task belongs to later, e.g. when it is found that a task takes
too long.