tests: Check that connecting lambdas with type changes works
This has been broken since commit 4bd7b5eb617f but the error actually
comes from the wrong implementation of

    active_source_first::sink::result_t;

which returns the result_t of source (!!) instead of sink. We did it
this way because it is impossible to determine the result type of sink
by _only_ looking at it.

Since then we figured out that we can get the result_t of a sink by
checking what it returns when called with result_t of source. The only
thing we missed was this wrong typedef.

Will be fixed in the following commit.