cmake: Use scoped dependencies in target_link_libraries
This takes care of boost library dependencies and relocateability of
built artifacts. We have to install the FindBoost.cmake module to our
cmake folder so that this works with older cmake versions (they don't
include the same version of this module). The installed package config
file now does the FindPackage stuff before including flexcore-targets.