cmake: Add FindBoost module from cmake v3.6
Has the nice property that it includes namespaced imported targets for
every component of boost. The older versions of this module do not have
this feature. This makes relocatable builds much easier as there are now
no hardcoded paths to system libraries.

The (BSD) license text has been substituted into the module file so
redistribution should be ok.