cmake: Use -pthread flag for threading
Some platforms (and boost) require this in addition to linking against
libpthread (due to the _REENTRANT define). This requires running

    SET(THREADS_PREFER_PTHREAD_FLAG ON)

before the call to find_package(Threads).

On some platforms (like Solaris) the thread library is integrated with
the C library and the FindThread.cmake module wrongly concludes that
nothing is needed to compile a pthread application. This leads to no
compile/link flags being added on the Threads::Threads target even when
the above variable is set - and boost decides that without the define
there will be no threading. This should be reported upstream to CMake.