Add microbenchmarks using google benchmark suite
I started to add microbenchmarks to test flexcore performance
against pure c++ implementations as baselines.
The benchmarks are meant to showcase the low performance overhead of flexcore
as well as allow us to test the effect of internal optimizations and prevent performance regressions.

The first benchmark tests the transformation of an std::vector<float> with a simple loop
vs flexcore map actions.