nodes: Add helper to create a new tree_base_node in tree
and convert existing add_node() functions to use the helper.

I don't believe anyone really uses this API (yet), but it is needed to
prevent a node from racing to get to its tree metadata in the
constructor before it is actually placed in the tree. We will overcome
this by first placing a tree_base_node with metadata in the tree, then
giving this metadata to the actual node we are constructing, and then
swapping the placeholder node with the desired one. This will require
creating the dummy node, hence the need for this helper in additional
places.