tests: Check whether nodes can access full_name in constructor
Although we did not want to make this behaviour available, it is
impossible to block people from checking their name in the constructor.
In thise case it is better to make it safe, so include tests that shows
that it currently doesn't work. These tests will be fixed in the next
commits.

Both tests fail, but for different reasons. When tree_base_node checks
its name, it's not part of the forest yet, so the assert in find_self()
fires. When owning_node checks its name, an owner_holder is already in
the forest, but the owner holder delegates all calls to the held owner,
which has not been set yet. So the unique_ptr assert fires.