nodes: Insert nodes into forest in 2 phases
This ensures that at any point in time (including the node constructor),
the metadata held within tree_base_node is valid, (i.e.
find_self(fg_->forest, *this) will not fire an assert). This is achieved
by:

 1. inserting a placeholder tree_base_node into forest,
 2. creating a unique_ptr<node> with the metadata of the node
    created in 1. outside the forest,
 3. and swapping the nodes created in 1. and 2.

The dynamic_cast could well be a static_cast but having it be dynamic
makes it easier to find mistakes in the logic. The same code is now used
for both make_child_impls (owning and leaf), and owner_holder is in
principle obsoleted but refactoring will be done separately.