forest_owner: Also use 2 phase insertion
This is the same code as used to insert nodes into forest later on, see
owning_base_node::make_child_impl. This has proven much simpler than
using the additional indirection of an owner_holder. With this, only one
of the constructors of owning_base_node is used, the other one can be
removed.