Add minimalistic benchmarks comparing ports and nodes against c++ functions
Currently the benchmark does not tell much.
The lambda, the pure port and the virtual function are all completly inlined.
only the complex code using the event_terminal takes real time.

It is quite interesting, that using Og instead of O0 with gcc makes a real difference in timings here.