Cleanup and Documentation of cyclecontrol api
moved definition of tick_rates to parallel_region.
User Code which passes tick_rates as parameters is going to use them with parallel_regions
and not with the scheduler directly.

moved conidition_pair to detail namespace as it is implemention only.

Add comments and asserts to cyclecontrol api.

made constructor if periodic task explicit.
This makes the intent of user code clear, when they pass a task
that this task is not a one shot, but periodic.