made connection constexpr
core connections are now constepxr and can be evaluated at compiletime

Add constexpr specifier to core functions and structs in connection.hpp
Provide an overload for const connection objects.
Until this the call operator would not work on const connections,
even if all members provided const operator()

Added concept check of operator>> to connect.
this allows to use the connect function with all checks instead of operator>>
when use of overloaded operators is not wanted.

Removed some obsolete documentations and dummy parameters in core/connection.hpp