trait variable templates _v are now bool instead of auto
when these traits are evaluated with more complex types gcc 5.2.0
falsly deduces the type of these variables to the tested type.
When the values are set to be bools, (which they are anyways)
the bug could not be reproduced.

The bug is fixed in gcc 5.2.1 but 5.2.0 is the default in
some versionf of the yocto project.