cmake: Replace uses of string(APPEND for cmake-3.3 compat
In order to support cmake 3.3, all uses of 'string(APPEND' (used only in
FindBoost.cmake which we imported from cmake 3.8?) are being replaced by
list(APPEND and then a final string(CONCAT. While not fully equivalent,
e.g. when ';' is in one of the list elements, this should not happen in
practice. It seems that the uses are actually restricted to error
messages and WIN32 anyway.